refactor: move custom hooks to their own folder
