+++ src/core/react/hooks/index.tsx
@@ -1,0 +1,1 @@
+export { useInterval } from './useInterval';
